Fix native ethernet (#37)
* Replace the default Ethernet library with the NativeEthernet library

The Teensy 4.1 requires a direct replacement for the normal
Ethernet.h library by the NativeEthernet library.

NativeEthernet is a builtin library in platformio, therefore no extra
installations are required.

* fix de-referencing null pointer

The locator struct is not handed over, but the transport_* callbacks
depend on it. This result in a  NULL pointer in `transport->args`.