timer: proper way to check POSIX timers support
The _POSIX_TIMERS macros can be used to determine the degree of a
host's support for the posix timers API.

The POSIX definition (2018 edition) says that:

_POSIX_TIMERS

The implementation supports timers. This symbol shall always
be set to the value 200809L.

It's defined in the <unistd.h> header.

Tested on esp32, pico and teensy4.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>