platform/core update for pico pico2 (#181)
* pico: switch to community supported platform/core and add pico2 support

As the develpment of pico/pico2 support on offical platformio had been
stalled for years, we have switched to platform/core contributed by
earlephilhower and maxgerhardt. It includes the latest toolchains gcc
14.3 and a lot of drivers/libraries.

The micro-ROS common.meta defines "-DRCUTILS_NO_64_ATOMIC=ON", which
enables atomic ops functions in micro-ros/rcutils. But pico-sdk has
build-in atomic ops, so they conflict and result in link errors. An
user meta is added to avoid the errors.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>

* ci: use the lastest ubuntu LTS 24.04

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>

* README: update for pico2 and more turorials

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>

* update default distro to kilted

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>

---------

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>